refactoring almost done
